[Trinity] Core/Entities: First step to 128 bit guids
    * Database fields storing full guid have been converted to BINARY(16)

+++ src/server/game/Battlegrounds/Bracket.cpp
@@ -114,7 +114,7 @@ void Bracket::SaveStats(SQLTransaction* trans)
     {
         case BRACKET_NEW:
             stmt = CharacterDatabase.GetPreparedStatement(CHAR_REP_CHARACTER_BRACKETS_STATS);
-            stmt->setUInt32(index++, m_owner.GetCounter());
+            stmt->setUInt64(index++, m_owner.GetCounter());
             stmt->setUInt8(index++, m_Type);
             stmt->setUInt16(index++, m_rating);
             stmt->setUInt16(index++, values[BRACKET_BEST]);
@@ -135,7 +135,7 @@ void Bracket::SaveStats(SQLTransaction* trans)
             stmt->setUInt32(index++, values[BRACKET_SEASON_WIN]);
             stmt->setUInt32(index++, values[BRACKET_WEEK_GAMES]);
             stmt->setUInt32(index++, values[BRACKET_WEEK_WIN]);
-            stmt->setUInt32(index++, m_owner.GetCounter());
+            stmt->setUInt64(index++, m_owner.GetCounter());
             stmt->setUInt8(index++, m_Type);
             break;
         default:
